+++ packages/create/src/generators/wc-lit-element/templates/_MyEl.js
@@ -4,8 +4,11 @@ export class <%= className %> extends LitElement {
   static get styles() {
     return css`
       :host {
+        --<%= tagName %>-text-color: #000;
+
         display: block;
         padding: 25px;
+        color: var(--<%= tagName %>-text-color);
       }
     `;
   }